Travis CI enablement
- Added test_requirements, test harness, and travis changes
- Made tempfile context manager importable
- Made context manager remeber it's current directory and return to it
