+++ .travis.yml
@@ -1,0 +1,13 @@
+language: python
+python:
+  - "2.7"
+before_install:
+  - sudo add-apt-repository ppa:texlive-backports/ppa -y
+  - sudo apt-get update -y
+  - sudo apt-get install libxml2-utils dvipng python-numpy python-scipy texlive texlive-latex-extra poppler-utils -y
+install:
+  - pip install -r requirements.txt --use-mirrors --allow-unverified PyXML
+  - pip install -r test_requirements.txt --use-mirrors
+script:  bash test.sh -q
+after_success:
+  coveralls
