+++ latex2edx/test/test_policy.py
@@ -1,20 +1,14 @@
 import os
 import re
-import contextlib
 import unittest
-import tempfile
-import shutil
 import json
 from path import path	# needs path.py
-import latex2edx as l2emod
-from latex2edx.main import latex2edx
 from StringIO import StringIO
 
-@contextlib.contextmanager
-def make_temp_directory():
-    temp_dir = tempfile.mkdtemp('l2etmp')
-    yield temp_dir
-    shutil.rmtree(temp_dir)
+import latex2edx as l2emod
+from latex2edx.main import latex2edx
+from latex2edx.test.util import make_temp_directory
+
 
 class TestPolicy(unittest.TestCase):
