+++ requirements.txt
@@ -1,7 +1,7 @@
 BeautifulSoup==3.2.0
 Cython==0.20.1
 Pillow==2.5.3
-PyXML==0.8.4 --allow-unverified PyXML
+PyXML==0.8.4
 -e git+https://github.com/mitocw/latex2dnd.git@master#egg=latex2dnd-master
 -e git+https://github.com/mitocw/latex2edx.git@master#egg=latex2edx-master
 lxml==3.3.3
