+++ test_requirements.txt
@@ -1,0 +1,7 @@
+nose
+coverage
+pep8
+pylint
+diff-cover
+pytest
+pyparsing
